♻️ Split book status page into sales and details tabs with unified edit
Query-synced tabs separate reporting (orders, sales channels, Plus stats)
from book management (metadata, listing settings, editions, purchase links).
Metadata and listing settings now display together read-only; an explicit
Edit mode saves with a dirty diff — chain tx only when on-chain fields
changed, POST /settings only when listing fields changed — and no longer
navigates away after saving.
