+++ i18n/locales/en.json
@@ -853,11 +853,19 @@
   "seo_titles.publish_nft_book": "Publish your new work",
   "seo.latest_books_title": "Latest Books",
   "seo.mint_nft_book_title": "Mint Liker Land NFT Book",
+  "status_page.book_details_title": "Book Details & Listing Settings",
+  "status_page.cancel_edit": "Cancel",
   "status_page.channel_ids_placeholder": "Channel ID(s), separated by commas (e.g. store01, store02)",
   "status_page.email_placeholder": "abc{at}example.com",
   "status_page.mark_complete": "Mark Complete",
+  "status_page.no_changes": "No changes to save",
   "status_page.search_placeholder": "Search...",
   "status_page.send_reminder_email": "Send Reminder Email",
+  "status_page.settings_saved": "Changes saved",
+  "status_page.tab_book_details": "Book Details",
+  "status_page.tab_sales_orders": "Sales & Orders",
+  "status_page.value_no": "No",
+  "status_page.value_yes": "Yes",
   "status_page.view_transaction": "View Transaction",
   "status_page.wallet_placeholder": "0x...",
   "status.completed": "Delivered",
