+++ i18n/locales/zh-TW.json
@@ -853,11 +853,19 @@
   "seo_titles.publish_nft_book": "出版您的新作",
   "seo.latest_books_title": "最新書籍",
   "seo.mint_nft_book_title": "鑄造 Liker Land NFT 書籍",
+  "status_page.book_details_title": "書籍資料與販售設定",
+  "status_page.cancel_edit": "取消",
   "status_page.channel_ids_placeholder": "頻道 ID，以逗號分隔（例如 ＠store01, ＠store02）",
   "status_page.email_placeholder": "abc{at}example.com",
   "status_page.mark_complete": "標記為完成",
+  "status_page.no_changes": "沒有變更需要儲存",
   "status_page.search_placeholder": "搜尋...",
   "status_page.send_reminder_email": "發送提醒電子郵件",
+  "status_page.settings_saved": "已儲存變更",
+  "status_page.tab_book_details": "書籍資料",
+  "status_page.tab_sales_orders": "銷售與訂單",
+  "status_page.value_no": "否",
+  "status_page.value_yes": "是",
   "status_page.view_transaction": "檢視交易",
   "status_page.wallet_placeholder": "0x...",
   "status.completed": "完成",
